# packages/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cocoon-SIT API",
    "description": "Cocoon-SIT backend implementing REST 202 + WebSocket realtime contracts, versioned prompt templates, audit artifacts, and worker-driven runtime execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/memory/{cocoon_id}": {
      "get": {
        "tags": [
          "memory"
        ],
        "summary": "List Memory",
        "operationId": "list_memory_api_v1_memory__cocoon_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoryChunkOut"
                  },
                  "title": "Response List Memory Api V1 Memory  Cocoon Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/{cocoon_id}/compact": {
      "post": {
        "tags": [
          "memory"
        ],
        "summary": "Compact Memory",
        "operationId": "compact_memory_api_v1_memory__cocoon_id__compact_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCompactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurableJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memory/{cocoon_id}/{memory_id}": {
      "delete": {
        "tags": [
          "memory"
        ],
        "summary": "Delete Memory",
        "operationId": "delete_memory_api_v1_memory__cocoon_id___memory_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          },
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryChunkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DurableJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cocoon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cocoon Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "lock_key": {
            "type": "string",
            "title": "Lock Key"
          },
          "payload_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload Json"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "worker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Name"
          },
          "error_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cocoon_id",
          "job_type",
          "status",
          "lock_key",
          "payload_json",
          "available_at",
          "started_at",
          "finished_at",
          "worker_name",
          "error_text"
        ],
        "title": "DurableJobOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryChunkOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "tags_json": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "summary",
          "content",
          "tags_json",
          "created_at"
        ],
        "title": "MemoryChunkOut"
      },
      "MemoryCompactionRequest": {
        "properties": {
          "before_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Message Id"
          }
        },
        "type": "object",
        "title": "MemoryCompactionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}